// Soe.Angular/TypeScript/Common/Models/CustomerInvoiceIODTO.ts
export class CustomerInvoiceIODTO implements ICustomerInvoiceIODTO {
    actorCompanyId: number;
    attestStateNames: string[];
    attestStates: number[];
    autoCreateCustomer: boolean;
    batchId: string;
    billingAddressAddress: string;
    billingAddressCity: string;
    billingAddressCO: string;
    billingAddressCountry: string;
    billingAddressName: string;
    billingAddressPostNr: string;
    billingType: number;
    billingTypeName: string;
    centRounding: number;
    claimAccountNr: string;
    claimAccountNrDim2: string;
    claimAccountNrDim3: string;
    claimAccountNrDim4: string;
    claimAccountNrDim5: string;
    claimAccountNrDim6: string;
    claimAccountNrSieDim1: string;
    claimAccountNrSieDim6: string;
    contractEndDay: number;
    contractEndMonth: number;
    contractEndYear: number;
    contractGroupDayInMonth: number;
    contractGroupDecription: string;
    contractGroupId: number;
    contractGroupInterval: number;
    contractGroupInvoiceRowText: string;
    contractGroupInvoiceTemplate: number;
    contractGroupInvoiceText: string;
    contractGroupName: string;
    contractGroupOrderTemplate: number;
    contractGroupPeriod: string;
    contractGroupPriceManagementName: string;
    contractNr: string;
    createAccountingInXE: boolean;
    created: Date;
    createdBy: string;
    createDeliveryAddressAsTextOnly: boolean;
    currency: string;
    currencyDate: Date;
    currencyId: number;
    currencyRate: number;
    customerExternalNr: string;
    customerExternaId: string;
    customerGlnNr: string;
    customerId: number;
    customerInvoiceHeadIOId: number;
    customerInvoiceNr: string;
    customerName: string;
    customerNr: string;
    customerOrgnr: string;
    debetInvoiceNr: string;
    deliveryAddressAddress: string;
    deliveryAddressCity: string;
    deliveryAddressCO: string;
    deliveryAddressCountry: string;
    deliveryAddressName: string;
    deliveryAddressPostNr: string;
    deliveryCondition: string;
    deliveryConditionId: number;
    deliveryDate: Date;
    deliveryType: string;
    deliveryTypeId: number;
    dueDate: Date;
    email: string;
    errorMessage: string;
    externalDescription: string;
    externalId: string;
    freightAmount: number;
    freightAmountCurrency: number;
    fullyPayed: boolean;
    import: boolean;
    importHeadType: number;
    internalDescription: string;
    invoiceDate: Date;
    invoiceDeliveryType: number;
    invoiceFee: number;
    invoiceFeeCurrency: number;
    invoiceHeadText: string;
    invoiceId: number;
    invoiceLabel: string;
    invoiceState: number;
    isClosed: boolean;
    isModified: boolean;
    isSelected: boolean;
    language: string;
    modified: Date;
    modifiedBy: string;
    nextContractPeriodDate: Date;
    nextContractPeriodValue: number;
    nextContractPeriodYear: number;
    note: string;
    nrOfProductRows: number;
    ocr: string;
    offerNr: string;
    orderNr: string;
    orderType: number;
    originStatus: number;
    originType: number;
    private _paidAmount: number;
    public get paidAmount(): number {
        return this._paidAmount;
    }
    public set paidAmount(value: number) {
        this._paidAmount = value;
    }
    paidAmountCurrency: number;
    paymentCondition: string;
    paymentConditionCode: string;
    paymentConditionId: number;
    paymentNr: string;
    priceListTypeId: number;
    projectId: number;
    projectNr: string;
    referenceOur: string;
    referenceYour: string;
    registrationType: number;
    remainingAmount: number;
    saleAccountNr: string;
    saleAccountNrDim2: string;
    saleAccountNrDim3: string;
    saleAccountNrDim4: string;
    saleAccountNrDim5: string;
    saleAccountNrDim6: string;
    saleAccountNrSieDim1: string;
    saleAccountNrSieDim6: string;
    seqNr: number;
    source: number;
    status: number;
    statusName: string;
    sumAmount: number;
    sumAmountCurrency: number;
    totalAmount: number;
    totalAmountCurrency: number;
    transferType: string;
    type: number;
    useFixedPriceArticle: boolean;
    vatAccountNr: string;
    vATAmount: number;
    vatAmount1: number;
    vatAmount2: number;
    vatAmount3: number;
    vATAmountCurrency: number;
    vatRate1: number;
    vatRate2: number;
    vatRate3: number;
    vatType: number;
    workingDescription: string;
    voucherDate: Date;
    voucherNr: string;
}